refactor(api): move inMemoryCache file and test
