+++ api/lib/infrastructure/caches/learning-content-cache.js
@@ -1,6 +1,6 @@
 import { config } from '../../../src/shared/config.js';
 import { DistributedCache } from '../../../src/shared/infrastructure/caches/DistributedCache.js';
-import { InMemoryCache } from './InMemoryCache.js';
+import { InMemoryCache } from '../../../src/shared/infrastructure/caches/InMemoryCache.js';
 import { LayeredCache } from './LayeredCache.js';
 import { RedisCache } from './RedisCache.js';
